# pkg/server/openapi.yaml
openapi: 3.1.0
info:
  title: native-ops daemon API
  version: set-at-runtime
  summary: Plan, deploy and inspect one Incus host from CI, an agent or a dashboard.
  description: |
    The native-ops daemon runs on the host it manages. Every change to the host is a **job** with a
    record; one change runs at a time (a second gets `409 busy`).

    **Deploys** are the normal way to change the host: push a protected `deploy-*` tag in the
    configuration repository and call `POST /v1/deploy` with it. The daemon reads that commit from the
    git server itself; nothing is uploaded, and the tag push is the approval. `POST /v1/deploy/plan`
    shows what a tag would change first; `GET /v1/deploys` says what the host runs now.

    **Roles** are ordered: viewer < planner < deployer < admin. Each operation names the least role it
    needs in `x-native-ops-role`. A token with a **scope** (for a fleet manager) may only call the
    operations marked `x-native-ops-scoped: true`, and only for instances its scope allows.

    Some operations exist only when the daemon is started with the flag in `x-native-ops-enabled-by`;
    otherwise they answer 404.

    Errors are `{"error": "<message>", "code": "<code>"}`. A failure to read the host is a generic
    `502 host_unavailable` that never echoes internal text.

    The same operations are available to AI agents as MCP tools at `POST /mcp` (see docs/api.md).
  license:
    name: MIT
    identifier: MIT
servers:
  - url: "{daemon}"
...
security:
  - bearer: []
...
paths:
...
  /v1/status:
    get:
      operationId: status
      tags: [host]
      summary: The host's state
      description: |
        Instances, data volumes, images, host metrics (load, memory, swap, pool disk, uptime) and
        warnings; with the edge and DNS configured, the edge's routes and certificates and the DNS
        records. Instance config is reported by key name only. Cached for a few seconds.
      x-native-ops-role: viewer
      responses:
        "200": { description: "The snapshot.", content: { application/json: { schema: { type: object } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/HostUnavailable" }
...
  /v1/jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: Jobs, newest first (without logs)
      description: A scoped token sees only jobs it started and jobs about instances its scope allows.
      x-native-ops-role: viewer
      x-native-ops-scoped: true
      parameters:
        - name: kind
          in: query
          description: A job kind, or a family of kinds ending in a colon (`instance:`).
          schema: { type: string, examples: [deploy, apply, "instance:", image:build, image:prune, edge:apply, backup, restore, dns:sync, daemon:upgrade] }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/JobStatus" } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        "200":
          description: The jobs.
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs: { type: array, items: { $ref: "#/components/schemas/Job" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/backups/restore:
    post:
      operationId: restore
      tags: [maintenance]
      summary: Restore a volume in place (keeping a copy) or as a new volume
      x-native-ops-role: admin
      x-native-ops-enabled-by: --enable-backup
      parameters:
        - { name: volume, in: query, required: true, schema: { type: string } }
        - { name: from, in: query, description: "The object key, or latest.", schema: { type: string, default: latest } }
        - { name: as, in: query, description: "Restore under this new volume name.", schema: { type: string } }
        - { name: force, in: query, description: "Stop the containers that mount the volume.", schema: { type: string, enum: ["1", "true"] } }
      requestBody: { $ref: "#/components/requestBodies/Tree" }
      responses:
        "202": { $ref: "#/components/responses/JobStarted" }
        "409": { $ref: "#/components/responses/Busy" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: An API token (nops_...), created by an admin (POST /v1/tokens or native-ops token create). The UI's session cookie also works, except at /mcp.
...
  requestBodies:
    Tree:
      required: true
      description: The configuration tree as a gzip-compressed tar (fleet.yml at its root, or in one top-level directory).
      content:
        application/gzip:
          schema: { type: string, contentEncoding: binary }
  responses:
    JobStarted:
      description: The job started. Location names it; follow it with GET /v1/jobs/{id}?wait=60.
      headers:
        Location: { schema: { type: string } }
      content:
        application/json:
          schema:
            type: object
            properties:
              job: { $ref: "#/components/schemas/Job" }
    BadRequest:
      description: The request is malformed (bad_request, bad_config).
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: No valid token (unauthorized). WWW-Authenticate says how to authenticate.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Busy:
      description: Another change is running on this host (busy); the message names its job. Wait for it and retry.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    HostUnavailable:
      description: The host could not be read (host_unavailable).
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error: { type: string, description: "What went wrong, for a person." }
        code:
          type: string
          description: Stable, for a program.
          examples: [bad_request, bad_config, unauthorized, forbidden, not_found, busy, tag_unresolved, git_unavailable, host_unavailable, plan_changed, blocked, stale, not_approvable, not_approved, approval_used, approval_expired, not_managed, internal]
...
    JobStatus:
      type: string
      enum: [running, succeeded, failed, interrupted]
      description: interrupted means the daemon stopped while the job ran; run it again (applies converge).
    Job:
      type: object
      required: [id, kind, status, actor, created]
      properties:
        id: { type: string, examples: [j-1791054271-18f384d8] }
        kind: { type: string, examples: [deploy, apply, "instance:put", image:build] }
        status: { $ref: "#/components/schemas/JobStatus" }
        actor: { type: string, description: "The token or user that started it." }
        sha: { type: string, description: "The commit (a deploy's, once resolved)." }
        service: { type: string, description: "What the job is about; for a deploy, the tag." }
        plan_hash: { type: string }
        created: { type: string, format: date-time }
        finished: { type: string, format: date-time }
        error: { type: string }
        result: { $ref: "#/components/schemas/DeployResult" }
        log: { type: string, description: "For deployers and admins; not in lists." }
        log_truncated: { type: boolean }
    DeployResult:
      type: string
      enum: [applied, no_changes, daemon_upgraded]
      description: How a deploy that succeeded ended. daemon_upgraded means it moved the daemon to fleet.yml's pin and resumed as a new deploy job.
...